requirements.txt
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.1.2
click==8.1.8
colorama==0.4.6
email-validator==2.1.0.post1
fastapi==0.115.12
h11==0.14.0
httptools==0.6.4
idna==3.10
jinja2==3.1.3
passlib==1.7.4
pydantic==2.11.3
pydantic_core==2.33.1
python-dotenv==1.1.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.2
reportlab==4.0.8
requests==2.31.0
sniffio==1.3.1
sqlalchemy==2.0.28
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.1
watchfiles==1.0.5
websockets==15.0.1
requests
